https://github.com/sequelize/sequelize/issues/9294
Z was used in the default provided by Sequelize, but trips up plain ole MSSQL DATETIME casting in an UPDATE so we can't make this a default.
This means that the app must know it's timezone and must cast to/from UTC intelligently OR we need to add an additional datatype (or subtype)
to deal with DATETIME2 and DATETIMEOFFSET properly.